UPD: processing improvements

+++ src/Telemetry.Domain/Entities/KpiRecord.cs
@@ -2,17 +2,17 @@
 
 public class KpiRecord
 {
-    public int RecordCount { get; set; } = 0;
+    public long RecordCount { get; set; } = 0;
     public double Speed { get; set; } = 0;
-    public int SpeedRecords { get; set; } = 0;
+    public long SpeedRecords { get; set; } = 0;
     public double OilTempC { get; set; } = 0;
-    public int OilRecords { get; set; } = 0;
+    public long OilRecords { get; set; } = 0;
     public double EngineRpm { get; set; } = 0;
-    public int EngineRecords { get; set; } = 0;
+    public long EngineRecords { get; set; } = 0;
     public double EmisionCO2 { get; set; } = 0;
-    public int Co2Records { get; set; } = 0;
-    public int CoolantHighTemperatureCount { get; set; } = 0;
+    public long Co2Records { get; set; } = 0;
+    public long CoolantHighTemperatureCount { get; set; } = 0;
     public double FuelConsumption { get; set; } = 0;
     public bool IsMoving { get; set; } = true;
-    public int StopsCount { get; set; } = 0;
+    public long StopsCount { get; set; } = 0;
 }
